requirements.txt
Django>=4.2
Pillow>=10.0
python-dotenv>=1.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
